Day 15 ongoing

+++ src/template/template.kt
@@ -5,12 +5,12 @@ import java.io.File
 
 fun main() {
 
-    fun part1(input: String, debug: Boolean = false): Int {
+    fun part1(input: String, debug: Boolean = false): Long {
 
         return 0
     }
 
-    fun part2(input: String, debug: Boolean = false): Int {
+    fun part2(input: String, debug: Boolean = false): Long {
 
 
         return 0
@@ -22,11 +22,11 @@ fun main() {
 
     val input = File("src", "Day05.txt").readText()
 
-    part1(testInput) test Pair(0, "test 1 part 1")
-    part1(input) test Pair(0, "part 1")
+    part1(testInput) test Pair(0L, "test 1 part 1")
+    part1(input) test Pair(0L, "part 1")
 
-    part2(testInput) test Pair(0, "test 2 part 2")
-    part2(input) test Pair(0, "part 2")
+    part2(testInput) test Pair(0L, "test 2 part 2")
+    part2(input) test Pair(0L, "part 2")
 
 
 }
